Software Validation from Datasets/Evaluation Tool/requirements.txt
numpy
pandas
pyarrow
PyYAML
scipy
soundfile
tqdm
matplotlib
